프로젝트 구조 설계 (#1) - patrick
