+++ src/main/java/com/aniwhere/domain/user/config/SuccessHandler.java
@@ -1,0 +1,25 @@
+package com.aniwhere.domain.user.config;
+
+import org.springframework.security.core.Authentication;
+import org.springframework.security.web.authentication.AuthenticationSuccessHandler;
+import org.springframework.security.web.savedrequest.HttpSessionRequestCache;
+import org.springframework.security.web.savedrequest.RequestCache;
+
+
+import jakarta.servlet.ServletException;
+import jakarta.servlet.http.HttpServletRequest;
+import jakarta.servlet.http.HttpServletResponse;
+import java.io.IOException;
+
+public class SuccessHandler implements AuthenticationSuccessHandler {
+
+    private RequestCache requestCache = new HttpSessionRequestCache();
+
+    @Override
+    public void onAuthenticationSuccess(HttpServletRequest request, HttpServletResponse response,
+                                        Authentication authentication) throws IOException, ServletException {
+        // SavedRequest 를 제거하고 루트 경로로 리디렉션
+        requestCache.removeRequest(request, response);
+        response.sendRedirect("/");
+    }
+}
