--- src/main/java/com/aniwhere/domain/user/info/api/ViewController.java
@@ -1,12 +1,0 @@
-package com.aniwhere.domain.user.info.api;
-
-import org.springframework.stereotype.Controller;
-import org.springframework.web.bind.annotation.GetMapping;
-import org.springframework.web.bind.annotation.RequestMapping;
-
-@Controller("userInfoViewController")
-@RequestMapping("/user")
-public class ViewController {
-
-
-}
