--- src/main/java/com/aniwhere/domain/user/info/controller/UserRestController.java
@@ -1,9 +1,0 @@
-package com.aniwhere.domain.user.info.controller;
-
-import org.springframework.web.bind.annotation.RequestMapping;
-import org.springframework.web.bind.annotation.RestController;
-
-@RestController("userRestController")
-@RequestMapping("/register")
-public class UserRestController {
-}
