--- src/main/java/com/aniwhere/domain/user/info/domain/User.java
@@ -1,7 +1,0 @@
-package com.aniwhere.domain.user.info.domain;
-
-import lombok.Data;
-
-@Data
-public class User {
-}
