--- src/main/java/com/aniwhere/domain/user/info/dto/UserDTO.java
@@ -1,7 +1,0 @@
-package com.aniwhere.domain.user.info.dto;
-
-import lombok.Data;
-
-@Data
-public class UserDTO {
-}
