--- src/main/java/com/aniwhere/domain/user/info/mapper/UserMapper.java
@@ -1,7 +1,0 @@
-package com.aniwhere.domain.user.info.mapper;
-
-import org.apache.ibatis.annotations.Mapper;
-
-@Mapper
-public class UserMapper {
-}
