--- src/main/java/com/aniwhere/domain/user/info/service/UserService.java
@@ -1,7 +1,0 @@
-package com.aniwhere.domain.user.info.service;
-
-import org.springframework.stereotype.Service;
-
-@Service
-public class UserService {
-}
