+++ src/main/java/com/aniwhere/domain/user/loginSession/controller/LoginController.java
@@ -1,0 +1,15 @@
+package com.aniwhere.domain.user.loginSession.controller;
+
+import org.springframework.stereotype.Controller;
+import org.springframework.web.bind.annotation.GetMapping;
+
+
+@Controller
+public class LoginController {
+
+    @GetMapping("/login")
+    public String loginPage() {
+        return "popup/login";
+    }
+
+}
