+++ src/main/java/com/aniwhere/domain/user/loginSession/domain/Login.java
@@ -1,0 +1,11 @@
+package com.aniwhere.domain.user.loginSession.domain;
+
+import lombok.Data;
+
+@Data
+public class Login {
+
+    private String username;
+    private String password;
+
+}
