+++ src/main/java/com/aniwhere/domain/user/loginSession/dto/LoginDTO.java
@@ -1,0 +1,11 @@
+package com.aniwhere.domain.user.loginSession.dto;
+
+import lombok.Data;
+
+@Data
+public class LoginDTO {
+
+    private String userName;
+    private String password;
+    private String role;
+}
