+++ src/main/java/com/aniwhere/domain/user/permission/controller/PermissionRestController.java
@@ -1,0 +1,9 @@
+package com.aniwhere.domain.user.permission.controller;
+
+import org.springframework.web.bind.annotation.RequestMapping;
+import org.springframework.web.bind.annotation.RestController;
+
+@RestController("permissionRestController")
+@RequestMapping("/perm")
+public class PermissionRestController {
+}
