+++ src/main/java/com/aniwhere/domain/user/permission/domain/Permission.java
@@ -1,0 +1,7 @@
+package com.aniwhere.domain.user.permission.domain;
+
+import lombok.Data;
+
+@Data
+public class Permission {
+}
