+++ src/main/java/com/aniwhere/domain/user/permission/dto/PermissionDTO.java
@@ -1,0 +1,7 @@
+package com.aniwhere.domain.user.permission.dto;
+
+import lombok.Data;
+
+@Data
+public class PermissionDTO {
+}
