+++ src/main/java/com/aniwhere/domain/user/permission/service/PermissionService.java
@@ -1,0 +1,7 @@
+package com.aniwhere.domain.user.permission.service;
+
+import org.springframework.stereotype.Service;
+
+@Service
+public class PermissionService {
+}
