+++ src/main/java/com/aniwhere/domain/user/register/domain/Join.java
@@ -1,0 +1,12 @@
+package com.aniwhere.domain.user.register.domain;
+
+import lombok.Data;
+
+@Data
+public class Join {
+
+    private String userName;
+    private String password;
+    private String email;
+    private String address;
+}
