--- src/main/java/com/aniwhere/domain/usermng/loginSession/domain/Login.java
@@ -1,11 +1,0 @@
-package com.aniwhere.domain.usermng.loginSession.domain;
-
-import lombok.Data;
-
-@Data
-public class Login {
-
-    private String username;
-    private String password;
-
-}
