--- src/main/java/com/aniwhere/domain/usermng/loginSession/service/CustomUserDetails.java
@@ -1,56 +1,0 @@
-package com.aniwhere.domain.usermng.loginSession.service;
-
-import org.springframework.security.core.GrantedAuthority;
-import org.springframework.security.core.authority.SimpleGrantedAuthority;
-import org.springframework.security.core.userdetails.UserDetails;
-
-import java.util.Collection;
-import java.util.Collections;
-
-public class CustomUserDetails implements UserDetails {
-
-    private final String username;
-    private final String password;
-    private final Collection<? extends GrantedAuthority> authorities;
-
-    public CustomUserDetails(String username, String password, String role) {
-        this.username = username;
-        this.password = password;
-        this.authorities = Collections.singletonList(new SimpleGrantedAuthority(role));
-    }
-
-    @Override
-    public Collection<? extends GrantedAuthority> getAuthorities() {
-        return authorities;
-    }
-
-    @Override
-    public String getPassword() {
-        return password;
-    }
-
-    @Override
-    public String getUsername() {
-        return username;
-    }
-
-    @Override
-    public boolean isAccountNonExpired() {
-        return true;
-    }
-
-    @Override
-    public boolean isAccountNonLocked() {
-        return true;
-    }
-
-    @Override
-    public boolean isCredentialsNonExpired() {
-        return true;
-    }
-
-    @Override
-    public boolean isEnabled() {
-        return true;
-    }
-}
