--- src/main/java/com/aniwhere/domain/usermng/loginSession/service/LoginService.java
@@ -1,33 +1,0 @@
-package com.aniwhere.domain.usermng.loginSession.service;
-
-import com.aniwhere.domain.usermng.loginSession.dto.LoginDTO;
-import com.aniwhere.domain.usermng.mapper.UserMngMapper;
-import org.springframework.beans.factory.annotation.Autowired;
-import org.springframework.security.core.userdetails.UserDetails;
-import org.springframework.security.core.userdetails.UserDetailsService;
-import org.springframework.security.core.userdetails.UsernameNotFoundException;
-import org.springframework.stereotype.Service;
-
-@Service
-public class LoginService implements UserDetailsService {
-
-    private final UserMngMapper userMngMapper;
-
-    @Autowired
-    public LoginService(UserMngMapper userMngMapper) {
-        this.userMngMapper = userMngMapper;
-    }
-
-    @Override
-    public UserDetails loadUserByUsername(String username) throws UsernameNotFoundException {
-        LoginDTO loginDTO = userMngMapper.findByUsername(username);
-        if (loginDTO == null) {
-            throw new UsernameNotFoundException("User not found");
-        }
-        return new CustomUserDetails(
-                loginDTO.getUserName(),
-                loginDTO.getPassword(),
-                loginDTO.getRole()
-        );
-    }
-}
