--- src/main/java/com/aniwhere/domain/usermng/permission/controller/PermissionRestController.java
@@ -1,9 +1,0 @@
-package com.aniwhere.domain.usermng.permission.controller;
-
-import org.springframework.web.bind.annotation.RequestMapping;
-import org.springframework.web.bind.annotation.RestController;
-
-@RestController("permissionRestController")
-@RequestMapping("/perm")
-public class PermissionRestController {
-}
