--- src/main/java/com/aniwhere/domain/usermng/permission/domain/Permission.java
@@ -1,7 +1,0 @@
-package com.aniwhere.domain.usermng.permission.domain;
-
-import lombok.Data;
-
-@Data
-public class Permission {
-}
