--- src/main/java/com/aniwhere/domain/usermng/permission/dto/PermissionDTO.java
@@ -1,7 +1,0 @@
-package com.aniwhere.domain.usermng.permission.dto;
-
-import lombok.Data;
-
-@Data
-public class PermissionDTO {
-}
