--- src/main/java/com/aniwhere/domain/usermng/permission/service/PermissionService.java
@@ -1,7 +1,0 @@
-package com.aniwhere.domain.usermng.permission.service;
-
-import org.springframework.stereotype.Service;
-
-@Service
-public class PermissionService {
-}
