--- src/main/java/com/aniwhere/domain/usermng/register/domain/Join.java
@@ -1,10 +1,0 @@
-package com.aniwhere.domain.usermng.register.domain;
-
-import lombok.Data;
-
-@Data
-public class Join {
-
-    private String userName;
-    private String password;
-}
