--- src/main/java/com/aniwhere/domain/usermng/register/dto/JoinDTO.java
@@ -1,11 +1,0 @@
-package com.aniwhere.domain.usermng.register.dto;
-
-import lombok.Data;
-
-@Data
-public class JoinDTO {
-
-    private String userName;
-    private String password;
-    private String role;
-}
