+++ src/main/resources/mybatis/mapper/usermng.xml
@@ -4,25 +4,25 @@
         PUBLIC "-//mybatis.org//DTD Mapper 3.0//EN"
         "http://mybatis.org/dtd/mybatis-3-mapper.dtd">
 
-<mapper namespace="com.aniwhere.domain.usermng.mapper.UserMngMapper">
+<mapper namespace="com.aniwhere.domain.user.mapper.UserMngMapper">
 
     <!-- 사용자 이름으로 사용자 정보를 찾는 SQL 쿼리 -->
-    <select id="findByUsername" resultType="com.aniwhere.domain.usermng.loginSession.dto.LoginDTO">
-        SELECT username, password, role
-        FROM users
+    <select id="findByUsername" resultType="com.aniwhere.domain.user.loginSession.dto.LoginDTO">
+        SELECT username, password, email, role
+        FROM Users
         WHERE username = #{username}
     </select>
 
     <!-- 사용자 정보를 추가하는 SQL 쿼리 -->
-    <insert id="insertUser" parameterType="com.aniwhere.domain.usermng.register.dto.JoinDTO">
-        INSERT INTO users (username, password, role)
-        VALUES (#{userName}, #{password}, #{role})
+    <insert id="insertUser" parameterType="com.aniwhere.domain.user.register.dto.JoinDTO">
+        INSERT INTO Users (username, password, email, address, role)
+        VALUES (#{userName}, #{password}, #{email}, #{address}, #{role})
     </insert>
 
     <!-- 사용자 이름으로 사용자 존재 여부를 확인하는 SQL 쿼리 -->
     <select id="checkByUserDatabase" resultType="boolean">
         SELECT IF(COUNT(*) > 0, true, false)
-        FROM users
+        FROM Users
         WHERE username = #{username}
     </select>
