+++ src/main/resources/static/js/course-create.js
@@ -1,0 +1,1 @@
+
