+++ src/main/resources/templates/aniwhere/route/course-create.html
@@ -26,17 +26,20 @@
         <div class="right-container">
             <div class="info-container">
                 <div><span class="title">경로 등록</span></div>
-                <form id="route-form">
-                    <div class="form-group">
-                        <label for="route-name">경로 이름</label>
-                        <input type="text" id="route-name" name="route-name" class="form-control" required>
-                    </div>
-                    
-                    <div class="form-group">
-                        <label for="route-description">경로 설명</label>
-                        <textarea id="route-description" name="route-description" class="form-control" required></textarea>
-                    </div>
-                </form>
+                <div>
+                    <form id="route-form">
+                        <div class="form-group">
+                            <label for="route-name">경로 이름</label>
+                            <input type="text" id="route-name" name="route-name" class="form-control" required>
+                        </div>
+                        
+                        <div class="form-group">
+                            <label for="route-description">경로 설명</label>
+                            <textarea id="route-description" name="route-description" class="form-control" required></textarea>
+                        </div>
+                    </form>
+                </div>
+                
                 <div class="directions-info"></div>
                 <div id="total-d" class="total-d"></div>
