+++ src/main/resources/templates/include/main-header.html
@@ -102,6 +102,16 @@
                         <input type="password" name="password" required="">
                         <label>비밀번호</label>
                     </div>
+                    <div class="user-box">
+                        <input type="text" name="email" required="">
+                        <label>이메일</label>
+                    </div>
+                    <!-- kakao addr api (js) -->
+                    <div class="user-box">
+                        <input type="text" name="address" required="">
+                        <label>주소</label>
+                    </div>
+                    
                     <div class="icon">
                         <button type="submit">
                             <i class="fa-solid fa-right-to-bracket"></i>
